Web Actions Tab: Factor out components
* More readable code with factored out components from main container
* Code only moved, no actual changes
