+++ misk/web/tabs/web-actions/src/components/WebActionsComponent.tsx
@@ -1,86 +1,34 @@
 import {
-  Button,
   Card,
   Classes,
   Collapse,
-  ControlGroup,
   H3,
   H5,
-  HTMLSelect,
   Icon,
-  InputGroup,
-  Intent,
-  Label,
   Menu,
   MenuItem,
   Spinner,
   Tag,
-  TextArea,
   Tooltip
 } from "@blueprintjs/core"
 import { IconNames } from "@blueprintjs/icons"
-import { CodePreContainer, FlexContainer } from "@misk/core"
-import {
-  onChangeFnCall,
-  onChangeToggleFnCall,
-  onClickFnCall,
-  simpleSelect
-} from "@misk/simpleredux"
-import { chain } from "lodash"
+import { FlexContainer } from "@misk/core"
+import { onChangeToggleFnCall, simpleSelect } from "@misk/simpleredux"
 import { HTTPMethod } from "http-method-enum"
+import { chain } from "lodash"
 import * as React from "react"
-import { connect } from "react-redux"
 import styled from "styled-components"
 import {
+  FilterWebActionsComponent,
+  SendRequestCollapseComponent
+} from "../components"
+import {
+  HTTPMethodIntent,
   IDispatchProps,
   IState,
   IWebActionInternal,
-  rootDispatcher,
-  rootSelectors,
   WebActionInternalLabel
 } from "../ducks"
-
-const tag = "WebActions"
-const filterTag = `${tag}::Filter`
-
-//TODO(adrw) upstream to @misk/core
-const HTTPMethodIntent: { [method in HTTPMethod]: Intent } = {
-  [HTTPMethod.CONNECT]: Intent.DANGER,
-  [HTTPMethod.DELETE]: Intent.DANGER,
-  [HTTPMethod.GET]: Intent.PRIMARY,
-  [HTTPMethod.HEAD]: Intent.WARNING,
-  [HTTPMethod.OPTIONS]: Intent.NONE,
-  [HTTPMethod.PATCH]: Intent.SUCCESS,
-  [HTTPMethod.POST]: Intent.SUCCESS,
-  [HTTPMethod.PUT]: Intent.SUCCESS,
-  [HTTPMethod.TRACE]: Intent.NONE
-}
-
-const HTTPStatusCodeIntent = (code: number) => {
-  if (200 <= code && code < 300) {
-    return Intent.SUCCESS
-  } else if (300 <= code && code < 400) {
-    return Intent.PRIMARY
-  } else if (400 <= code && code < 500) {
-    return Intent.WARNING
-  } else if (500 <= code && code < 600) {
-    return Intent.DANGER
-  } else {
-    return Intent.NONE
-  }
-}
-
-const HTTPMethodDispatch: any = (props: IDispatchProps) => ({
-  [HTTPMethod.CONNECT]: props.simpleNetworkGet,
-  [HTTPMethod.DELETE]: props.simpleNetworkDelete,
-  [HTTPMethod.GET]: props.simpleNetworkGet,
-  [HTTPMethod.HEAD]: props.simpleNetworkHead,
-  [HTTPMethod.OPTIONS]: props.simpleNetworkGet,
-  [HTTPMethod.PATCH]: props.simpleNetworkPatch,
-  [HTTPMethod.POST]: props.simpleNetworkPost,
-  [HTTPMethod.PUT]: props.simpleNetworkPut,
-  [HTTPMethod.TRACE]: props.simpleNetworkGet
-})
 
 const FloatLeft = styled.span`
   float: left;
@@ -216,189 +164,10 @@ const MethodTag = (props: { method: HTTPMethod }) => (
 )
 
 /**
- * Collapse wrapped Send a Request form for each Web Action card
- */
-const SendRequestCollapse = (
-  props: { action: IWebActionInternal } & IState & IDispatchProps
-) => {
-  // Determine if Send Request form for the Web Action should be open
-  const isOpen =
-    simpleSelect(
-      props.simpleForm,
-      `${tag}::${props.action.pathPattern}::Request`,
-      "data"
-    ) || false
-  const url = simpleSelect(
-    props.simpleForm,
-    `${tag}::${props.action.pathPattern}::URL`,
-    "data"
-  )
-  // Pre-populate the URL field with the action path pattern on open of request form
-  if (isOpen && !url) {
-    props.simpleFormInput(
-      `${tag}::${props.action.pathPattern}::URL`,
-      props.action.pathPattern
-    )
-  }
-  const actionTag = `${tag}::${props.action.function}::${
-    props.action.pathPattern
-  }`
-  const method: HTTPMethod =
-    simpleSelect(props.simpleForm, `${actionTag}::Method`, "data") ||
-    props.action.dispatchMechanism.reverse()[0]
-  const methodHasBody =
-    method === HTTPMethod.PATCH ||
-    method === HTTPMethod.POST ||
-    method === HTTPMethod.PUT
-  return (
-    <Collapse isOpen={isOpen}>
-      <InputGroup
-        defaultValue={props.action.pathPattern}
-        onChange={onChangeFnCall(props.simpleFormInput, `${actionTag}::URL`)}
-        placeholder={
-          "Request URL: absolute ( http://your.url.com/to/send/a/request/to/ ) or internal service endpoint ( /service/web/action )"
-        }
-        type={"url"}
-      />
-      <Collapse isOpen={methodHasBody}>
-        <TextArea
-          fill={true}
-          onChange={onChangeFnCall(props.simpleFormInput, `${actionTag}::Body`)}
-          placeholder={
-            "Request Body (JSON or Text).\nDrag bottom right corner of text area input to expand."
-          }
-        />
-      </Collapse>
-      <ControlGroup>
-        <HTMLSelect
-          large={true}
-          onChange={onChangeFnCall(
-            props.simpleFormInput,
-            `${actionTag}::Method`
-          )}
-          options={props.action.dispatchMechanism.sort()}
-          value={method}
-        />
-        <Button
-          large={true}
-          onClick={onClickFnCall(
-            HTTPMethodDispatch(props)[method],
-            `${actionTag}::Response`,
-            url,
-            simpleSelect(props.simpleForm, `${actionTag}::Body`, "data")
-          )}
-          intent={HTTPMethodIntent[method]}
-          loading={simpleSelect(
-            props.simpleNetwork,
-            `${actionTag}::Response`,
-            "loading"
-          )}
-          text={"Submit"}
-        />
-      </ControlGroup>
-      <Label>
-        Request <Tag>{url}</Tag>
-      </Label>
-      <Collapse
-        isOpen={simpleSelect(props.simpleForm, `${actionTag}::Body`, "data")}
-      >
-        <CodePreContainer>
-          {JSON.stringify(
-            simpleSelect(props.simpleForm, `${actionTag}::Body`, "data"),
-            null,
-            2
-          )}
-        </CodePreContainer>
-      </Collapse>
-      <Collapse
-        isOpen={simpleSelect(
-          props.simpleNetwork,
-          `${actionTag}::Response`,
-          "status"
-        )}
-      >
-        <Label>
-          Response{" "}
-          <Tag
-            intent={HTTPStatusCodeIntent(
-              simpleSelect(
-                props.simpleNetwork,
-                `${actionTag}::Response`,
-                "status"
-              )[0]
-            )}
-          >
-            {(
-              simpleSelect(
-                props.simpleNetwork,
-                `${actionTag}::Response`,
-                "status"
-              ) || []
-            ).join(" ")}
-          </Tag>{" "}
-          <Tag
-            intent={Intent.NONE}
-            onClick={onChangeToggleFnCall(
-              props.simpleFormToggle,
-              `${actionTag}::ButtonRawResponse`,
-              props.simpleForm
-            )}
-          >
-            <span>
-              Raw Response{" "}
-              {simpleSelect(
-                props.simpleForm,
-                `${actionTag}::ButtonRawResponse`,
-                "data"
-              ) ? (
-                <Icon icon={IconNames.CARET_DOWN} />
-              ) : (
-                <Icon icon={IconNames.CARET_RIGHT} />
-              )}
-            </span>
-          </Tag>
-        </Label>
-      </Collapse>
-      <Collapse
-        isOpen={simpleSelect(
-          props.simpleNetwork,
-          `${actionTag}::Response`,
-          "data"
-        )}
-      >
-        <CodePreContainer>
-          {JSON.stringify(
-            simpleSelect(props.simpleNetwork, `${actionTag}::Response`, "data"),
-            null,
-            2
-          )}
-        </CodePreContainer>
-      </Collapse>
-      <Collapse
-        isOpen={simpleSelect(
-          props.simpleForm,
-          `${actionTag}::ButtonRawResponse`,
-          "data"
-        )}
-      >
-        <Label>Raw Network Redux State</Label>
-        <CodePreContainer>
-          {JSON.stringify(
-            simpleSelect(props.simpleNetwork, `${actionTag}::Response`),
-            null,
-            2
-          )}
-        </CodePreContainer>
-      </Collapse>
-    </Collapse>
-  )
-}
-
-/**
  * Web Action Card rendered for each bound Web Action
  */
 const WebAction = (
-  props: { action: IWebActionInternal } & IState & IDispatchProps
+  props: { action: IWebActionInternal; tag: string } & IState & IDispatchProps
 ) => (
   <div>
     <Card>
@@ -451,7 +220,7 @@ const WebAction = (
             <MetadataCollapse
               content={"Application Interceptors"}
               label={`(${props.action.applicationInterceptors.length})`}
-              tag={`${tag}::${
+              tag={`${props.tag}::${
                 props.action.pathPattern
               }::ApplicationInterceptors`}
               {...props}
@@ -463,7 +232,9 @@ const WebAction = (
             <MetadataCollapse
               content={"Network Interceptors"}
               label={`(${props.action.networkInterceptors.length})`}
-              tag={`${tag}::${props.action.pathPattern}::NetworkInterceptors`}
+              tag={`${props.tag}::${
+                props.action.pathPattern
+              }::NetworkInterceptors`}
               {...props}
             >
               {props.action.networkInterceptors.map(i => (
@@ -473,7 +244,7 @@ const WebAction = (
             <MetadataCollapse
               content={"Send a Request"}
               label={""}
-              tag={`${tag}::${props.action.pathPattern}::Request`}
+              tag={`${props.tag}::${props.action.pathPattern}::Request`}
               {...props}
             >
               <span />
@@ -481,7 +252,7 @@ const WebAction = (
           </MetadataMenu>
         </Column>
       </FlexContainer>
-      <SendRequestCollapse {...props} />
+      <SendRequestCollapseComponent {...props} />
     </Card>
     <br />
   </div>
@@ -491,7 +262,8 @@ const WebAction = (
  * Loops over bound Web Actions array (metadata) and renders Web Action card
  */
 const WebActions = (
-  props: { metadata: IWebActionInternal[] } & IState & IDispatchProps
+  props: { metadata: IWebActionInternal[]; tag: string } & IState &
+    IDispatchProps
 ) => {
   if (props.metadata) {
     return (
@@ -504,33 +276,6 @@ const WebActions = (
   } else {
     return <Spinner />
   }
-}
-
-const FilterWebActions = (props: IState & IDispatchProps) => {
-  const FilterSelectOptions = [...Object.keys(WebActionInternalLabel)]
-  return (
-    <div>
-      <ControlGroup fill={true}>
-        <HTMLSelect
-          large={true}
-          onChange={onChangeFnCall(
-            props.simpleFormInput,
-            `${filterTag}::HTMLSelect`
-          )}
-          options={FilterSelectOptions}
-        />
-        <InputGroup
-          large={true}
-          onChange={onChangeFnCall(
-            props.simpleFormInput,
-            `${filterTag}::Input`
-          )}
-          placeholder={"Filter Web Actions"}
-        />
-        {/* TODO(adrw) Use multiselect to do autocomplete filters */}
-      </ControlGroup>
-    </div>
-  )
 }
 
 /**
@@ -578,8 +323,11 @@ const SkeletonWebActions = () => (
   </Card>
 )
 
-const WebActionsContainer = (props: IState & IDispatchProps) => {
+export const WebActionsComponent = (
+  props: IState & IDispatchProps & { tag: string }
+) => {
   const metadata = simpleSelect(props.webActions, "metadata")
+  const filterTag = `${props.tag}::Filter`
   if (metadata.length > 0) {
     const filterKey =
       WebActionInternalLabel[
@@ -601,7 +349,7 @@ const WebActionsContainer = (props: IState & IDispatchProps) => {
       .value()
     return (
       <div>
-        <FilterWebActions {...props} />
+        <FilterWebActionsComponent {...props} />
         <WebActions
           metadata={filteredMetadata as IWebActionInternal[]}
           {...props}
@@ -612,7 +360,7 @@ const WebActionsContainer = (props: IState & IDispatchProps) => {
     // Displays mock of 5 Web Action cards which fill in when data is available
     return (
       <div>
-        <FilterWebActions {...props} />
+        <FilterWebActionsComponent {...props} />
         <SkeletonWebActions />
         <br />
         <SkeletonWebActions />
@@ -626,10 +374,3 @@ const WebActionsContainer = (props: IState & IDispatchProps) => {
     )
   }
 }
-
-const mapStateToProps = (state: IState) => rootSelectors(state)
-
-export default connect(
-  mapStateToProps,
-  rootDispatcher
-)(WebActionsContainer)
