+++ misk/web/tabs/web-actions/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./FilterWebActionsComponent"
+export * from "./SendRequestCollapseComponent"
+export * from "./WebActionsComponent"
