+++ misk/web/tabs/web-actions/src/containers/index.ts
@@ -1,2 +1,1 @@
 export { default as TabContainer } from "./TabContainer"
-export { default as WebActionsContainer } from "./WebActionsContainer"
